Update index (3).html
So, I am working on fixing the navbar's sign-in and sign-up buttons. Still working on the color scheme on the site but I think I found one. In this patch, I fixed the logo that was drifting too far to the right and added the columns to the home page. Right now, I still have to fix the text formatting and footer links to display correctly 👍